remove unused code paths
