+++ scrubber/__init__.py
@@ -1,6 +1,3 @@
-from . import transform
-from . import geom
-from . import cli
 from . import core
 from . import storage
 from . import common
@@ -8,13 +5,10 @@ from .protonate import AcidBaseConjugator
 from .protonate import Tautomerizer
 from .ringfix import fix_rings
 from .core import Scrub
-from .core import gen3d
+from .geometry import gen3d
 from .storage import SMIMolSupplierWrapper
 
 __all__ = [
-    "transform",
-    "geom",
-    "cli",
     "core",
     "storage",
     "common",
@@ -23,4 +17,5 @@ __all__ = [
     "fix_rings",
     "Scrub",
     "gen3d",
+    "SMIMolSupplierWrapper",
 ]
