+++ scrubber/common.py
@@ -1,71 +1,4 @@
-import os
-
 from rdkit import Chem
-
-# this is the subdirectory name (with respect to the path of this file) where
-# the the default datafiles are stored
-DATA_DIR = "data"
-DATA_PATH = os.path.join(os.path.dirname(__file__), DATA_DIR)
-
-
-__all__ = ["DATA_PATH", "ScrubberBase", "UniqueMoleculeContainer"]
-
-
-class ScrubberBase():
-    """abstract class with the basic functionality that most (all?) Scrubber
-    classes should have.
-    A class method provide access to the class init defaults by returning the
-    class `__init__` signature with default values as a dictionary
-    `{arg:default_value}`.
-
-    Inheriting classes should not run the base class __init__() method, but
-    they must implement the `_stop_at_defaults` argument (preferably as the
-    last one) to guarantee the `get_defaults` method works properly.
-    >>> class ScrubberNewClass(ScrubberClass):
-    ...
-    ...     [ class implementation here ]
-    ...
-    # to present init options to a driver class:
-    >>> scrubber_class_options = ScrubberNewClass.get_defaults()
-    >>> scrubber_class_options['key'] = False
-    >>> sc = ScrubberClassNew(**scrubber_class_options)
-    """
-    def __init__(self, arguments=None, _stop_at_defaults: bool = False):
-        """generic implementation of the init;
-        inheriting classes will implement their own args, but keep the
-        __stop_at_defaults flag to support the get_defaults() class method and
-        add the clause below to allow early termination.
-
-        This method will create a disposable instance from which the defaults
-        will be extractedcore.
-        """
-        # print("GOT HERE, TOO")
-        self.options = {}
-        if _stop_at_defaults:
-            return
-
-    @classmethod
-    def get_defaults(cls):
-        """method to return the default values of init options"""
-        return cls(_stop_at_defaults=True).__dict__
-
-    def get_datafile(self, fname: str) -> str:
-        """helper function to provide the full-path of data files available in
-        the Scrubber module default
-        location ("scrubber/data", defined by the DATA_DIR variable)"""
-        fullpath = os.path.join(DATA_PATH, fname)
-        if not os.path.exists(fullpath):
-            msg = (
-                "ERROR: the specified filename [%s] "
-                "does not exist in the data "
-                "directory [%s]" % (fname, DATA_PATH),
-            )
-            raise ValueError(msg)
-        return fullpath
-
-    def _build_opts_dict(self):
-        """build and populate the self.options dictionary"""
-        self.options = self.__dict__.copy()
 
 
 class UniqueMoleculeContainer():
@@ -431,40 +364,3 @@ def mol2smi(mol, chirality=True):
     return Chem.MolToSmiles(
         Chem.RemoveHs(mol), isomericSmiles=chirality, canonical=True
     )
-
-
-def copy_mol_properties(
-    mol_source,
-    mol_dest,
-    strict: bool = False,
-    include_name: bool = True,
-    exclude: list = None,
-):
-    """convenience function to copy properties from a molecule to another; if
-    strict mode is requested, only molecules in the source molecule will be
-    kept in the destination molecule; an optional list of excluded properties
-    can be provided
-
-    computed properties are not copied
-    """
-    exclude = exclude or []
-    input_properties = mol_source.GetPropsAsDict()
-    # print("PROCESSING HTESE PROPERTIES", input_properties)
-    for p, v in input_properties.items():
-        if "_scrubber_tmp_" in p:
-            continue
-        if p in exclude:
-            continue
-        if type(v) == int:
-            mol_dest.SetIntProp(p, v)
-        elif type(v) == bool:
-            mol_dest.SetBoolProp(p, v)
-        elif type(v) == float:
-            mol_dest.SetDoubleProp(p, v)
-        elif type(v) == str:
-            mol_dest.SetProp(p, v)
-    if include_name:
-        if not mol_source.HasProp("_Name"):
-            return
-        name = mol_source.GetProp("_Name")
-        mol_dest.SetProp("_Name", name)
