--- scrubber/geom/__init__.py
@@ -1,2 +1,0 @@
-
-# in theory the ring transformation should happen here...
